move packages into the internal package
