+++ internal/bot/bot.go
