+++ internal/bot/doc.go
