+++ internal/bot/help.go
